x86/msr-index: Remove dependency on <linux/bits.h>
The x86 <asm/msr-index.h> header is (mistakenly) exported as a UAPI
header in 4.9.  Adding the inclusion of <linux/bits.h> stops that
working entirely as that header isn't exported.

This will likely be resolved upstream by un-exporting
<asm/msr-index.h>.  However, to avoid build regressions for user-space
in stretch, we should keep it working as a UAPI header.

Remove the inclusion of <linux/bits.h> and open-code BIT() for the
recently converted definitions.  There are other bits defined with
BIT_ULL() which won't work in user-space, but that's not a regression.

Gbp-Pq: Topic debian
Gbp-Pq: Name x86-msr-index-remove-dependency-on-linux-bits.h.patch