x86/msr-index: Remove dependency on <linux/bits.h>
The x86 <asm/msr-index.h> header is (mistakenly) exported as a UAPI
header in 4.9.  Adding the inclusion of <linux/bits.h> stops that
working entirely as that header isn't exported.

This has been resolved upstream by un-exporting <asm/msr-index.h>.
However, to avoid build regressions for user-space in stretch, we
should keep it working as a UAPI header.

Make the inclusion of <linux/bits.h> conditional on __KERNEL__.
Although the bit macros defined with BIT() and BIT_ULL() won't work in
user-space, that's not a regression.

Gbp-Pq: Topic debian
Gbp-Pq: Name x86-msr-index-remove-dependency-on-linux-bits.h.patch