inet: Avoid ABI change for IP ID hash change
"inet: switch IP ID generator to siphash" adds a new member to struct
netns_ipv4.  Since this is embedded in struct net, it changes the
offsets of all the following members.  However struct net itself is
not embedded in anything, and is always allocated by built-in code.
So move the new member to the end of struct net, and hide it from
genksyms.

Also hide the added element and member from modules, as they won't be
able to rely on their being present until we bump ABI.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name inet-avoid-abi-change-for-ip-id-hash-change.patch