drm/i915: Avoid ABI change for CVE-2019-0155
Commit 3e977ac6179b "drm/i915: Prevent writing into a read-only object
via a GGTT mmap" adds a new member to struct drm_vma_offset_node, a
DRM core structure.  This structure is embedded in some other DRM
structures, and doesn't seem to have any padding, so adding an extra
member anywhere in it is an ABI change.

However, the vm_node member is a structure with many padding bits free
after some bitfields.  Since the new member is a single flag, we can
replace it with a 1-bit bitfield in that structure.  We still need to
hide this from genksyms.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name drm-i915-avoid-abi-change-for-cve-2019-0155.patch