Prevent building uImage with missing mkimage
Bug-Debian: https://bugs.debian.org/708094
Forwarded: not-needed

On some powerpc platforms, CONFIG_DEFAULT_UIMAGE is selected automatically,
which leads to uImage being built automatically with mkimage. This tool is not
available in linux's build-dependencies, and the file is not strictly
necessary, so we are omitting this step in the build process, Debian-specific.

Gbp-Pq: Topic debian
Gbp-Pq: Name powerpcspe-omit-uimage.patch